mejoras en parte autmatica, mejora en UI
se mejoro lo funcional en calculo automatico y su UI

+++ app/src/main/java/com/example/kleberstevendiazcoello/ui/ViewHolder/AdapaterAutoHolder.java
@@ -38,7 +38,7 @@ public class AdapaterAutoHolder extends RecyclerView.Adapter<AdapaterAutoHolder.
         holder.comida.setText(arrayList.get(position).getFoodName());
        // holder.cantidad.setText(arrayList.get(position).getCantidad());
         //holder.Carbohidratos.setText(arrayList.get(position).getCalorias());
-        int total = (Integer.parseInt(arrayList.get(position).getCalorias()))*(Integer.parseInt(arrayList.get(position).getCantidad()));
+        float total = (Float.parseFloat(arrayList.get(position).getCalorias()))*(Integer.parseInt(arrayList.get(position).getCantidad()));
         holder.Carbohidratos.setText(String.valueOf(total));
     }
